requirements.txt
pandas==1.3.1
PyYAML==6.0
rich==12.6.0
SQLAlchemy==1.4.23
Telethon==1.25.4
